rename zendmm -> zend_mm
