+++ ext/zend_test/zend_mm_observer.stub.php
